+++ migrations/2.0_deploy_chain_storage.js
@@ -7,7 +7,8 @@ module.exports = (deployer, network, accounts) => {
 
   return deployer.deploy(
     ChainStorageArtifact,
-    config.Chain.blocksPerPhase,
+    config.Chain.blocksPerPropose,
+    config.Chain.blocksPerReveal,
     config.Chain.minimumStakingTokenPercentage,
     false,
     options,