+++ test/integration/TestEnableVerifiers.js
@@ -13,7 +13,8 @@ const {
 } = require('../helpers/deployers');
 
 const verifiersCount = 3;
-const phaseDuration = 5 * verifiersCount * 2;
+const proposePhaseDuration = 5 * verifiersCount * 2;
+const revealPhaseDuration = 5 * verifiersCount * 2;
 const requirePercentOfTokens = 70;
 
 contract('Chain: testing enabled/disabled verifiers', (accounts) => {
@@ -27,7 +28,7 @@ contract('Chain: testing enabled/disabled verifiers', (accounts) => {
 
   before(async () => {
     ministroChain = await deployChain(
-      accounts[0], verifiersAddr, phaseDuration,
+      accounts[0], verifiersAddr, proposePhaseDuration, revealPhaseDuration,
       requirePercentOfTokens, true,
     );
 
@@ -49,11 +50,11 @@ contract('Chain: testing enabled/disabled verifiers', (accounts) => {
 
     describe('when we are in propose phase', async () => {
       before(async () => {
-        await mineUntilPropose(phaseDuration);
+        await mineUntilPropose(proposePhaseDuration, revealPhaseDuration);
       });
 
       it('should NOT be possible to propose', async () => {
-        const blockHeight = await getBlockHeight(phaseDuration);
+        const blockHeight = await getBlockHeight(proposePhaseDuration, revealPhaseDuration);
         const awaits = [];
         for (let i = 0; i < verifiersCount; i += 1) {
           awaits.push(ministroChain.propose(
@@ -80,11 +81,11 @@ contract('Chain: testing enabled/disabled verifiers', (accounts) => {
 
         describe('when we are in propose phase', async () => {
           before(async () => {
-            await mineUntilPropose(phaseDuration);
+            await mineUntilPropose(proposePhaseDuration, revealPhaseDuration);
           });
 
           it('should be possible to propose', async () => {
-            const blockHeigght = await getBlockHeight(phaseDuration);
+            const blockHeigght = await getBlockHeight(proposePhaseDuration, revealPhaseDuration);
             const awaits = [];
             for (let i = 0; i < verifiersCount; i += 1) {
               awaits.push(ministroChain.propose(
@@ -109,7 +110,7 @@ contract('Chain: testing enabled/disabled verifiers', (accounts) => {
 
             describe('they should be able to finish election and reveal vote', async () => {
               before(async () => {
-                await mineUntilReveal(phaseDuration);
+                await mineUntilReveal(proposePhaseDuration, revealPhaseDuration);
               });
 
               it('should be possible to reveal', async () => {