+++ test/integration/TestUnusualCases.js
@@ -15,7 +15,8 @@ contract('Chain - unusual cases', (accounts) => {
 
   // that testRPC creates 1 block per tx, if you will get revert because of invalid phase,
   // then increase phase duration
-  const phaseDuration = verifiersCount * 5;
+  const proposePhaseDuration = verifiersCount * 5;
+  const revealPhaseDuration = verifiersCount * 5;
   const requirePercentOfTokens = 70;
 
   const proposalsObj = createProposals(verifiersCount, accounts);
@@ -25,22 +26,22 @@ contract('Chain - unusual cases', (accounts) => {
 
   before(async () => {
     ministroChain = await deployChain(
-      accounts[0], verifiersAddr, phaseDuration,
+      accounts[0], verifiersAddr, proposePhaseDuration, revealPhaseDuration,
       requirePercentOfTokens, true,
     );
 
     const blindedProposalCheck = await ministroChain.createProof(proposals[0], secrets[0]);
     assert.strictEqual(blindedProposals[0], blindedProposalCheck, 'Hasher do not produce same output as blockchain keccak256');
 
-    await mineUntilReveal(phaseDuration);
+    await mineUntilReveal(proposePhaseDuration, revealPhaseDuration);
   });
 
   describe('when all verifiers made a proposal', async () => {
     let blockHeigh;
 
     before(async () => {
-      await mineUntilPropose(phaseDuration);
-      const blockHeight = await getBlockHeight(phaseDuration);
+      await mineUntilPropose(proposePhaseDuration, revealPhaseDuration);
+      const blockHeight = await getBlockHeight(proposePhaseDuration, revealPhaseDuration);
 
       // start with i=1 so we have one "clean" option to propose for test purposes
       const awaits = [];
@@ -53,23 +54,25 @@ contract('Chain - unusual cases', (accounts) => {
       }
       await Promise.all(awaits);
 
-      blockHeigh = await ministroChain.getBlockHeight(phaseDuration);
+      blockHeigh = await ministroChain.getBlockHeight(proposePhaseDuration, revealPhaseDuration);
     });
 
     describe('when noone reveal', async () => {
       before(async () => {
-        await mineUntilReveal(phaseDuration);
-        await mineUntilPropose(phaseDuration);
+        await mineUntilReveal(proposePhaseDuration, revealPhaseDuration);
+        await mineUntilPropose(proposePhaseDuration, revealPhaseDuration);
       });
 
       it('should be next exlection cycle', async () => {
-        const blockHeigh2 = await ministroChain.getBlockHeight(phaseDuration);
+        const blockHeigh2 =
+          await ministroChain.getBlockHeight(proposePhaseDuration, revealPhaseDuration);
+
         assert(BigNumber(blockHeigh).plus(1).eq(blockHeigh2), 'invalid block height');
       });
 
 
       it('should be possible to propose again (because of new cycle)', async () => {
-        const blockHeight = await getBlockHeight(phaseDuration);
+        const blockHeight = await getBlockHeight(proposePhaseDuration, revealPhaseDuration);
         const awaits = [];
         for (let i = 1; i < verifiersCount; i += 1) {
           awaits.push(ministroChain.propose(
@@ -84,17 +87,20 @@ contract('Chain - unusual cases', (accounts) => {
 
       describe('when we enter NEXT reveal phase', async () => {
         before(async () => {
-          const blockHeight = await getBlockHeight(phaseDuration);
+          const blockHeight = await getBlockHeight(proposePhaseDuration, revealPhaseDuration);
+
           await ministroChain.propose(
             blindedProposals[0],
             blockHeight,
             { from: verifiersAddr[0] },
           );
-          await mineUntilReveal(phaseDuration);
+          await mineUntilReveal(proposePhaseDuration, revealPhaseDuration);
         });
 
         it('should still be next exlection cycle', async () => {
-          const blockHeigh2 = await ministroChain.getBlockHeight(phaseDuration);
+          const blockHeigh2 =
+            await ministroChain.getBlockHeight(proposePhaseDuration, revealPhaseDuration);
+
           assert(BigNumber(blockHeigh).plus(1).eq(blockHeigh2), 'invalid block height');
         });