+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+###Changed
+- Each phase has its own duration
+
 ## [0.5.1] - 2020-01-27
 ### Changed
 - when updating contract via `ContractRegistry`, old `Chain` contract is not killed